모든 엔티티 작성 완료 1 ( 코딩 하시다 안되는 부분이 있을 수 있으니 안되는 부분이 나오면 공유 부탁드립니다 )

+++ src/main/java/com/gcf/spring/repository/BannerOneRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.gcf.spring.entity.BannerOne;
 
-public interface BannerOneRepository extends JpaRepository<BannerOne, Long> {
+public interface BannerOneRepository extends JpaRepository<BannerOne, Integer> {
 
 }
